:art: (home) Refactor CTASection
Rename files for clarity

Move SpeakersGraphics to its own component

Update CTASale/SpeakersCall layout

Import SpeakersGraphics to CTASale/SpeakersCall

Adjust CTAApp spacing
