+++ src/views/HomePageView/CTASection/CTAApp/index.js
@@ -1,0 +1,1 @@
+export { default } from './CTAApp';
