+++ src/views/HomePageView/CTASection/CTASale/index.js
@@ -1,0 +1,1 @@
+export { default } from './CTASale';
