+++ src/views/HomePageView/CTASection/CTASpeakersCall/index.js
@@ -1,0 +1,1 @@
+export { default } from './CTASpeakersCall';
