--- src/views/HomePageView/CTASection/Sale/index.js
@@ -1,1 +1,0 @@
-export { default } from './Sale';
