--- src/views/HomePageView/CTASection/SpeakersCall/index.js
@@ -1,1 +1,0 @@
-export { default } from './SpeakersCall';
