+++ src/views/HomePageView/CTASection/SpeakersGraphics/index.js
@@ -1,0 +1,1 @@
+export { default } from './SpeakersGraphics';
